- fixed: MD2 vertex indices were signed shorts but should be unsigned. - fixed a few warnings from Visual C++ 2015.

+++ src/win32/fb_d3d9.cpp
@@ -3277,8 +3277,8 @@ void D3DFB::FillSimplePoly(FTexture *texture, FVector2 *points, int npoints,
 		return;
 	}
 
-	cosrot = cos(rot);
-	sinrot = sin(rot);
+	cosrot = (float)cos(rot);
+	sinrot = (float)sin(rot);
 
 	CheckQuadBatch(npoints - 2, npoints);
 	quad = &QuadExtra[QuadBatchPos];
